pyproject.toml
[project]
name = "arcane"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.2",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "apscheduler==3.11.0",
    "arrow==1.3.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "blessed==1.21.0",
    "bs4>=0.0.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "django==5.2.1",
    "django-apscheduler==0.7.0",
    "django-picklefield==3.3",
    "django-q2==1.8.0",
    "django-role-permissions==3.2.0",
    "faiss-cpu==1.11.0",
    "frozenlist==1.6.0",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.62",
    "langchain-openai==0.3.18",
    "langchain-text-splitters==0.3.8",
    "langchain-xai==0.2.4",
    "langsmith==0.3.42",
    "marshmallow==3.26.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "numpy==2.2.6",
    "openai==1.82.0",
    "orjson==3.10.18",
    "packaging==24.2",
    "propcache==0.3.1",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pypdf==5.5.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "redis==3.5.3",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sqlparse==0.5.3",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "types-python-dateutil==2.9.0.20250516",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'settings.py', 'manage.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'
